chore: test-case refinement
