+++ src/components/Pages/Header/Header.test.tsx
@@ -1,13 +1,7 @@
 // Modules
-import { useFeatureFlag } from '@/hooks/useFeatureFlag';
 import { cleanup, render, screen } from '@testing-library/react';
-import { MockInstance, afterEach, describe, expect, it } from 'vitest';
+import { afterEach, describe, expect, it } from 'vitest';
 import Header from './Header';
-
-// Mock feature flag hook
-vi.mock('@/hooks/useFeatureFlag', () => ({
-  useFeatureFlag: vi.fn(),
-}));
 
 // Render
 const renderComponent = () => render(<Header />);
@@ -15,7 +9,6 @@ const renderComponent = () => render(<Header />);
 // Suite
 describe('Header Component', () => {
   beforeEach(() => {
-    (useFeatureFlag as unknown as MockInstance).mockReturnValue(true);
     renderComponent();
   });
 
@@ -66,15 +59,5 @@ describe('Header Component', () => {
     renderComponent();
     const icon = await screen.findByTestId('header-accessibility-icon');
     expect(icon).not.toBeNull();
-  });
-
-  it('should render the fall-back under-development page when the page is not-available', () => {
-    (useFeatureFlag as unknown as MockInstance).mockReturnValue(false);
-
-    cleanup();
-    renderComponent();
-
-    const underDevelopmentPage = screen.queryAllByTestId('underdevelopment-page')[0];
-    expect(underDevelopmentPage).toBeTruthy();
   });
 });
